pyproject.toml
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[project]
name = "VTCBench"
version = "0.1.0"
description = "Needle-In-A-Haystack revisited from VLM's Perspective"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "jsonargparse[signatures]>=4.35.0,<4.43.0",
    "openai",
    "tenacity>=9.0.0",
    "tiktoken",
    "transformers>=4.47.1,<=4.57.1,!=4.57.0,!=4.52.0",
    "sentencepiece",
    "pillow",
    "deocr[playwright,pymupdf]>=0.3.0",
    "rouge-score>=0.1.2",
]

# for local development
# [tool.uv.sources]
# deocr = { path = "../DeOCR", editable = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "openpyxl>=3.1.5",
    "seaborn>=0.13.2",
]
